+++ pickly-service/build.gradle
@@ -30,7 +30,7 @@ dependencies {
     // p6spy
     implementation 'com.github.gavlyukovskiy:p6spy-spring-boot-starter:1.8.0'
 
-    runtimeOnly 'org.postgresql:postgresql:42.5.3'
+    runtimeOnly 'org.postgresql:postgresql:42.3.0'
     annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
@@ -38,7 +38,8 @@ dependencies {
     testImplementation 'org.testcontainers:testcontainers:1.16.0'
     testImplementation 'org.testcontainers:jdbc:1.16.0'
     testImplementation 'org.testcontainers:junit-jupiter:1.16.2'
-    testImplementation 'org.testcontainers:postgresql:1.16.0'
+    testImplementation "org.testcontainers:postgresql:1.17.6"
+
 }
 
 test {
