+++ pickly-service/src/test/groovy/org/pickly/service/friend/service/FriendServiceSpec.groovy
@@ -10,12 +10,18 @@ import org.pickly.service.member.entity.Member
 import org.pickly.service.member.entity.Password
 import org.pickly.service.member.repository.interfaces.MemberRepository
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.transaction.annotation.Transactional
+import org.testcontainers.junit.jupiter.Testcontainers
 import spock.lang.Specification
 
+@Transactional
 @SpringBootTest
-@AutoConfigureMockMvc
+@Testcontainers
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@ActiveProfiles("test")
 class FriendServiceSpec extends Specification {
 
     @Autowired
@@ -26,7 +32,6 @@ class FriendServiceSpec extends Specification {
 
     @Autowired
     private MemberRepository memberRepository
-
 
     @BeforeEach
     void setup() {
