+++ pickly-service/src/test/groovy/org/pickly/service/member/service/MemberServiceSpec.groovy
@@ -7,18 +7,22 @@ import org.pickly.service.category.CategoryFactory
 import org.pickly.service.category.repository.interfaces.CategoryRepository
 import org.pickly.service.friend.service.interfaces.FriendService
 import org.pickly.service.member.MemberFactory
-import org.pickly.service.member.entity.Member
-import org.pickly.service.member.entity.Password
 import org.pickly.service.member.repository.interfaces.MemberRepository
 import org.pickly.service.member.service.dto.MemberProfileUpdateDTO
 import org.pickly.service.member.service.interfaces.MemberService
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.test.context.ActiveProfiles
+import org.springframework.transaction.annotation.Transactional
+import org.testcontainers.junit.jupiter.Testcontainers
 import spock.lang.Specification
 
+@Transactional
 @SpringBootTest
-@AutoConfigureMockMvc
+@Testcontainers
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@ActiveProfiles("test")
 class MemberServiceSpec extends Specification {
 
     @Autowired
